Improving charts styles for adding to the dashboard.
