--- just-do-it/src/components/BasicCard/BasicCard.css
@@ -1,12 +1,0 @@
-.card-hero {
-    width: 350px;
-    height: 370px;
-    border-radius: 5px;
-    background-color: #E5E5E5;
-}
-
-.card-hero__title {
-    font-size: 16px;
-    font-weight: 500;
-    font-family: sans-serif;
-}
